requirements.txt
fastapi==0.103.2
uvicorn==0.24.0.post1
pydantic_settings==2.1.0
dnspython==2.4.2
aiokafka==0.9.0
asyncio==3.4.3
mongoengine==0.27.0
stringcase==1.2.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
devtools==0.12.2
pymongo==4.3.3
mongoengine==0.27.0
slowapi==0.1.9
pillow==10.0.1
python-multipart==0.0.20
boto3==1.36.13